Former-commit-id: 28747f2597472f0d43fd28a41f95df393b307ae5

+++ backend/netcha/src/main/java/com/netcha/member/data/MemberRepository.java
@@ -2,7 +2,7 @@ package com.netcha.member.data;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface MemberRepository extends CrudRepository<Member, Long> {
+public interface MemberRepository extends CrudRepository<Member, Integer> {
 	Member findByUserId(String userId);
 
 }
